Move classes from test_canvas to newly created integration/canvaas folder
